traycer/vlm-service/requirements.txt
fastapi>=0.115
uvicorn>=0.34
pydantic>=2.0
Pillow>=10.0
google-genai>=1.0
openai>=1.0
